fixed test issue and moved msw into setuptest file
